refactor(workflow): change workflow definition from factory style to function style
This commit removes the dependency on the `defineWorkflow` function throughout
the workflow system and changes the approach to use standalone async functions
instead. The main benefits of this change are:

- Simplified workflow definition using standard async functions
- More idiomatic TypeScript code with better type safety
- Improved testability of workflow functions
- Better support for serverless execution environments
- Reduced bundle size by removing the factory function overhead

Changes include:
- Updating all documentation examples to use the new function style
- Modifying the workflow editor to support the new format
- Updating seed files to align with the new approach
- Adapting runtime to work with standalone functions
- Improving workflow parser to detect function-style workflows
